Refactor series component/header
